pyproject.toml
[project]
name = "nebius-connect"
description = "Nebius AI connector for Apache Spark™"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "pyspark[connect]==3.5.*",
    "grpcio>=1.66.1",
    "setuptools>=75.1.0; python_version>='3.12'",
]
classifiers = [
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'License :: OSI Approved :: Apache Software License',
    'Operating System :: OS Independent',
]

[project.readme]
file = "README.md"
content-type = "text/markdown"
